feat: expose stylus IIR config schema

+++ EGoTouchService/Solvers/StylusSolver/shared/CoorIIRProcess.hpp
@@ -13,17 +13,17 @@ public:
 
     // ── IIR coefficient selection (GetIIRCoef equivalent) ──
     // In-band mode params (edge NOT active)
-    uint8_t m_coefLowInBand = 12;      // asa[0xA5E]
-    uint8_t m_coefHighInBand = 6;    // asa[0xA5F]
-    uint8_t m_speedTholdInBand = 20;  // 0x14
+    int32_t m_coefLowInBand = 12;     // asa[0xA5E]
+    int32_t m_coefHighInBand = 6;     // asa[0xA5F]
+    int32_t m_speedTholdInBand = 20;  // 0x14
 
     // Edge mode params (edge active)
-    uint8_t m_coefLowEdge = 6;     // asa[0xA5C]
-    uint8_t m_coefHighEdge = 18;   // asa[0xA5D]
-    uint8_t m_speedTholdEdge = 10; // 0x0A
+    int32_t m_coefLowEdge = 6;        // asa[0xA5C]
+    int32_t m_coefHighEdge = 18;      // asa[0xA5D]
+    int32_t m_speedTholdEdge = 10;    // 0x0A
 
-    int m_speedMax = 140;  // 0xCD — speed value at which high coef is fully engaged
-    uint8_t m_maxCoef = 32;   // asa[0xA60] — denominator in IIR formula
+    int32_t m_speedMax = 140;  // 0xCD — speed value at which high coef is fully engaged
+    int32_t m_maxCoef = 32;    // asa[0xA60] — denominator in IIR formula
 
     // ── Output ──
     uint16_t m_currentCoef = 10;
@@ -109,9 +109,9 @@ private:
     int m_counterA = 0;
 
     inline uint16_t SelectCoef(int speedValue, bool writingMode, bool peakOnEdge) const {
-        uint8_t coefLow;
-        uint8_t coefHigh;
-        int speedThreshold;
+        int32_t coefLow;
+        int32_t coefHigh;
+        int32_t speedThreshold;
 
         if (!writingMode) {
             // Hover (no pressure) uses In-band params
@@ -127,8 +127,8 @@ private:
 
         // Peak-on-edge overrides coefficients
         if (peakOnEdge) {
-            coefHigh = static_cast<uint8_t>(m_coefHighInBand >> 1); // 16 >> 1 = 8
-            coefLow = static_cast<uint8_t>(m_coefLowInBand >> 1);   // 2 >> 1 = 1
+            coefHigh = m_coefHighInBand >> 1; // 16 >> 1 = 8
+            coefLow = m_coefLowInBand >> 1;   // 2 >> 1 = 1
         }
 
         uint16_t selected;
